move pine64.py board under pine64 vendor dir
